Add files via upload
The project developed during the university course (Instituto Superior Técnico Lisboa) - Energy Services. The objective was to investigate machine learning capabilities in providing energy services like the analysis of electricity demand. The code was developed in Python language in the Google Colab notebook.
The available input data was 2 years of electricity usage in the building and the weather data.